backend/requirements.txt
flask==2.2.5
werkzeug==2.2.3
flask-cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.3.0
pytest==7.0.1
pytest-cov==3.0.0
pytest-mock==3.7.0
